what.c  	32/64 bit casts. system.h	32/64 bit casts. slib.c  	32/64 bit casts. randomBits.c	32/64 bit casts.
bk: 37dbddc4jKMP5BmyYcwNkub8wJJ86g

+++ src/port/randomBits.c
@@ -27,7 +27,7 @@ randomBits(char *buf)
 		u32 x, y;
 
 		gettimeofday(&tv, NULL);
-		x = (u32)sbrk(0) ^ (u32)&a;
+		x = (u32)(long)sbrk(0) ^ (u32)(long)&a;
 		y = ((u32)getpid() << 16) + (u32)getuid();
 		y ^= tv.tv_usec;
